Extract external contract code
